requirements.txt
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.0.0
nltk>=3.5
pandas>=1.1.0
Pillow>=8.0.0
tqdm>=4.50.0
seaborn>=0.11.0
matplotlib>=3.3.0
gensim>=4.0.0
GPUtil>=1.4.0
scikit-learn>=0.24.0
psutil>=5.8.0